Added /routes/shedules file and moved the schedules routes in. Stripped DB code from main index.js
